define content items for history
